server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
httpx>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
jinja2>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
aiofiles>=23.0.0
starlette>=0.36.0
aiosqlite>=0.20.0
itsdangerous>=2.1.0
